// types.ts
export interface TestModifier {
  testID?: string;
}
// ...
export type KSpacing =
  | '0.25rem' // 4px
  | '0.5rem' // 8px
  | '0.75rem' // 12px
  | '1rem' // 16px
  | '1.25rem' // 20px
  | '1.5rem' // 24px
  | '2rem' // 32px
  | '2.5rem' // 40px
  | '3rem' // 48px
  | '4rem' // 64px
  | '5rem' // 80px
  | '6rem' // 96px
  | '8rem' // 128px
  | '10rem' // 160px
  | '12rem' // 192px
  | '14rem' // 234px
  | '16rem'; // 256px
// ...
export type KVectorIconsProvider =
  | 'Droppii'
  | 'DroppiiNew'
  | 'Entypo'
  | 'EvilsIcons'
  | 'Feather'
  | 'FontAwesome'
  | 'Foundation'
  | 'Ionicons'
  | 'MaterialIcons'
  | 'MaterialCommunityIcons'
  | 'Octicons'
  | 'Zocial'
  | 'SimpleLineIcons';
// ...
export interface KButtonIconProps {
  vectorName?: string; // for vector icon
  vectorProvider?: KVectorIconsProvider; // for vector icon
  imageSource?: string | number; // for image icon,
  tintColor?: string;
  size?: number;
}
// ...
export interface KTabBarButtonProps extends TestModifier {
  label: string;
  icon?: KButtonIconProps;
}

export type KTabBarType = 'underline' | 'solid' | 'outline';

export type KTabBarTheme = 'primary' | 'secondary' | 'success' | 'warning' | 'danger' | 'dark';

export interface KTabBarProps extends TestModifier {
  theme?: KTabBarTheme;
  type: KTabBarType;
  buttons: KTabBarButtonProps[];
  initialIndex?: number;
  iconAlignment?: 'top' | 'left' | 'bottom' | 'right';
  containerStyle?: StyleProp<ViewStyle>;
  paddingH?: KSpacing | 0;
  onChanged: (index: number) => void;
}